Analyze differential glycan trait means and heterogeneity

A high-level wrapper that computes glycan trait abundance and heterogeneity scores from GPSM data and tests for differences between two experimental groups.

The function internally:

Computes site- and protein-level glycan heterogeneity scores.

Performs permutation-based or Welch's t-test group comparisons.

analyze_hscore_changes(
  gpsm,
  from,
  motifs = NULL,
  meta,
  group_col,
  group_levels,
  B = 1000,
  min_samples = 3,
  method = c("permutation", "t.test")
)

Arguments

gpsm

A GPSM table imported by read_pGlyco3_gpsm or read_decipher_gpsm.

from

Character string specifying the glycan representation. One of "pGlyco3" or "decipher".

motifs

Optional glycan motif definitions.

meta

Sample metadata table.

group_col

Column name in meta defining the comparison groups.

group_levels

Length-2 character vector specifying the two groups to compare.

B

Integer; number of random label permutations used when `method = "permutation"`. Default is 1000.

min_samples

Minimum number of non-missing samples required in each group.

method

Statistical test used for group comparison. One of `"permutation"` or `"t.test"`. Default is `"permutation"`.

Value

A data frame containing differential analysis results for all glycan traits.

Examples

gpsm <- readRDS(system.file("extdata", "gpsm_toyexample.rds", package = "glycoTraitR"))
meta <- readRDS(system.file("extdata", "meta_toyexample.rds", package = "glycoTraitR"))

res <- analyze_hscore_changes(
  gpsm = gpsm,
  from = "pGlyco3",
  meta = meta,
  group_col = "Diagnosis",
  group_levels = c("Normal", "Symptomatic"),
  B = 5
)
#> Compute heterogeneity scores at each site and protein
#>   |                                                  | 0 % ~calculating    |+                                                 | 1 % ~04s            |++                                                | 2 % ~04s            |++                                                | 4 % ~03s            |+++                                               | 5 % ~03s            |++++                                              | 6 % ~03s            |++++                                              | 7 % ~03s            |+++++                                             | 8 % ~03s            |+++++                                             | 10% ~03s            |++++++                                            | 11% ~03s            |+++++++                                           | 12% ~03s            |+++++++                                           | 13% ~03s            |++++++++                                          | 14% ~03s            |++++++++                                          | 16% ~03s            |+++++++++                                         | 17% ~03s            |++++++++++                                        | 18% ~03s            |++++++++++                                        | 19% ~03s            |+++++++++++                                       | 20% ~03s            |+++++++++++                                       | 22% ~03s            |++++++++++++                                      | 23% ~03s            |+++++++++++++                                     | 24% ~03s            |+++++++++++++                                     | 25% ~03s            |++++++++++++++                                    | 27% ~03s            |++++++++++++++                                    | 28% ~02s            |+++++++++++++++                                   | 29% ~02s            |++++++++++++++++                                  | 30% ~02s            |++++++++++++++++                                  | 31% ~02s            |+++++++++++++++++                                 | 33% ~02s            |+++++++++++++++++                                 | 34% ~02s            |++++++++++++++++++                                | 35% ~02s            |+++++++++++++++++++                               | 36% ~02s            |+++++++++++++++++++                               | 37% ~02s            |++++++++++++++++++++                              | 39% ~02s            |++++++++++++++++++++                              | 40% ~02s            |+++++++++++++++++++++                             | 41% ~02s            |++++++++++++++++++++++                            | 42% ~02s            |++++++++++++++++++++++                            | 43% ~02s            |+++++++++++++++++++++++                           | 45% ~02s            |+++++++++++++++++++++++                           | 46% ~02s            |++++++++++++++++++++++++                          | 47% ~02s            |+++++++++++++++++++++++++                         | 48% ~02s            |+++++++++++++++++++++++++                         | 49% ~02s            |++++++++++++++++++++++++++                        | 51% ~02s            |++++++++++++++++++++++++++                        | 52% ~02s            |+++++++++++++++++++++++++++                       | 53% ~02s            |++++++++++++++++++++++++++++                      | 54% ~02s            |++++++++++++++++++++++++++++                      | 55% ~01s            |+++++++++++++++++++++++++++++                     | 57% ~01s            |+++++++++++++++++++++++++++++                     | 58% ~01s            |++++++++++++++++++++++++++++++                    | 59% ~01s            |+++++++++++++++++++++++++++++++                   | 60% ~01s            |+++++++++++++++++++++++++++++++                   | 61% ~01s            |++++++++++++++++++++++++++++++++                  | 63% ~01s            |++++++++++++++++++++++++++++++++                  | 64% ~01s            |+++++++++++++++++++++++++++++++++                 | 65% ~01s            |++++++++++++++++++++++++++++++++++                | 66% ~01s            |++++++++++++++++++++++++++++++++++                | 67% ~01s            |+++++++++++++++++++++++++++++++++++               | 69% ~01s            |+++++++++++++++++++++++++++++++++++               | 70% ~01s            |++++++++++++++++++++++++++++++++++++              | 71% ~01s            |+++++++++++++++++++++++++++++++++++++             | 72% ~01s            |+++++++++++++++++++++++++++++++++++++             | 73% ~01s            |++++++++++++++++++++++++++++++++++++++            | 75% ~01s            |++++++++++++++++++++++++++++++++++++++            | 76% ~01s            |+++++++++++++++++++++++++++++++++++++++           | 77% ~01s            |++++++++++++++++++++++++++++++++++++++++          | 78% ~01s            |++++++++++++++++++++++++++++++++++++++++          | 80% ~01s            |+++++++++++++++++++++++++++++++++++++++++         | 81% ~01s            |+++++++++++++++++++++++++++++++++++++++++         | 82% ~01s            |++++++++++++++++++++++++++++++++++++++++++        | 83% ~01s            |+++++++++++++++++++++++++++++++++++++++++++       | 84% ~01s            |+++++++++++++++++++++++++++++++++++++++++++       | 86% ~00s            |++++++++++++++++++++++++++++++++++++++++++++      | 87% ~00s            |++++++++++++++++++++++++++++++++++++++++++++      | 88% ~00s            |+++++++++++++++++++++++++++++++++++++++++++++     | 89% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++    | 90% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++    | 92% ~00s            |+++++++++++++++++++++++++++++++++++++++++++++++   | 93% ~00s            |+++++++++++++++++++++++++++++++++++++++++++++++   | 94% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++++  | 95% ~00s            |+++++++++++++++++++++++++++++++++++++++++++++++++ | 96% ~00s            |+++++++++++++++++++++++++++++++++++++++++++++++++ | 98% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++++++| 99% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++++++| 100% elapsed=03s  
#> Test heterogeneity change at site
#>   |                                                  | 0 % ~calculating    |+                                                 | 1 % ~03s            |++                                                | 2 % ~03s            |++                                                | 3 % ~02s            |+++                                               | 4 % ~02s            |+++                                               | 5 % ~02s            |++++                                              | 6 % ~02s            |++++                                              | 7 % ~02s            |+++++                                             | 8 % ~02s            |+++++                                             | 9 % ~02s            |++++++                                            | 10% ~02s            |++++++                                            | 11% ~02s            |+++++++                                           | 12% ~02s            |+++++++                                           | 13% ~02s            |++++++++                                          | 14% ~02s            |++++++++                                          | 15% ~01s            |+++++++++                                         | 16% ~01s            |+++++++++                                         | 17% ~01s            |++++++++++                                        | 18% ~01s            |++++++++++                                        | 19% ~01s            |+++++++++++                                       | 20% ~01s            |+++++++++++                                       | 21% ~01s            |++++++++++++                                      | 22% ~01s            |++++++++++++                                      | 23% ~01s            |+++++++++++++                                     | 24% ~01s            |+++++++++++++                                     | 25% ~01s            |++++++++++++++                                    | 26% ~01s            |++++++++++++++                                    | 27% ~01s            |+++++++++++++++                                   | 28% ~01s            |+++++++++++++++                                   | 29% ~01s            |++++++++++++++++                                  | 30% ~01s            |++++++++++++++++                                  | 31% ~01s            |+++++++++++++++++                                 | 32% ~01s            |+++++++++++++++++                                 | 33% ~01s            |++++++++++++++++++                                | 34% ~01s            |++++++++++++++++++                                | 35% ~01s            |+++++++++++++++++++                               | 36% ~01s            |+++++++++++++++++++                               | 37% ~01s            |++++++++++++++++++++                              | 38% ~01s            |++++++++++++++++++++                              | 39% ~01s            |+++++++++++++++++++++                             | 40% ~01s            |+++++++++++++++++++++                             | 41% ~01s            |++++++++++++++++++++++                            | 42% ~01s            |++++++++++++++++++++++                            | 43% ~01s            |+++++++++++++++++++++++                           | 44% ~01s            |+++++++++++++++++++++++                           | 45% ~01s            |++++++++++++++++++++++++                          | 46% ~01s            |++++++++++++++++++++++++                          | 47% ~01s            |+++++++++++++++++++++++++                         | 48% ~01s            |+++++++++++++++++++++++++                         | 49% ~01s            |++++++++++++++++++++++++++                        | 51% ~01s            |++++++++++++++++++++++++++                        | 52% ~01s            |+++++++++++++++++++++++++++                       | 53% ~01s            |+++++++++++++++++++++++++++                       | 54% ~01s            |++++++++++++++++++++++++++++                      | 55% ~01s            |++++++++++++++++++++++++++++                      | 56% ~01s            |+++++++++++++++++++++++++++++                     | 57% ~01s            |+++++++++++++++++++++++++++++                     | 58% ~01s            |++++++++++++++++++++++++++++++                    | 59% ~01s            |++++++++++++++++++++++++++++++                    | 60% ~01s            |+++++++++++++++++++++++++++++++                   | 61% ~01s            |+++++++++++++++++++++++++++++++                   | 62% ~01s            |++++++++++++++++++++++++++++++++                  | 63% ~01s            |++++++++++++++++++++++++++++++++                  | 64% ~01s            |+++++++++++++++++++++++++++++++++                 | 65% ~01s            |+++++++++++++++++++++++++++++++++                 | 66% ~01s            |++++++++++++++++++++++++++++++++++                | 67% ~01s            |++++++++++++++++++++++++++++++++++                | 68% ~01s            |+++++++++++++++++++++++++++++++++++               | 69% ~01s            |+++++++++++++++++++++++++++++++++++               | 70% ~01s            |++++++++++++++++++++++++++++++++++++              | 71% ~00s            |++++++++++++++++++++++++++++++++++++              | 72% ~00s            |+++++++++++++++++++++++++++++++++++++             | 73% ~00s            |+++++++++++++++++++++++++++++++++++++             | 74% ~00s            |++++++++++++++++++++++++++++++++++++++            | 75% ~00s            |++++++++++++++++++++++++++++++++++++++            | 76% ~00s            |+++++++++++++++++++++++++++++++++++++++           | 77% ~00s            |+++++++++++++++++++++++++++++++++++++++           | 78% ~00s            |++++++++++++++++++++++++++++++++++++++++          | 79% ~00s            |++++++++++++++++++++++++++++++++++++++++          | 80% ~00s            |+++++++++++++++++++++++++++++++++++++++++         | 81% ~00s            |+++++++++++++++++++++++++++++++++++++++++         | 82% ~00s            |++++++++++++++++++++++++++++++++++++++++++        | 83% ~00s            |++++++++++++++++++++++++++++++++++++++++++        | 84% ~00s            |+++++++++++++++++++++++++++++++++++++++++++       | 85% ~00s            |+++++++++++++++++++++++++++++++++++++++++++       | 86% ~00s            |++++++++++++++++++++++++++++++++++++++++++++      | 87% ~00s            |++++++++++++++++++++++++++++++++++++++++++++      | 88% ~00s            |+++++++++++++++++++++++++++++++++++++++++++++     | 89% ~00s            |+++++++++++++++++++++++++++++++++++++++++++++     | 90% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++    | 91% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++    | 92% ~00s            |+++++++++++++++++++++++++++++++++++++++++++++++   | 93% ~00s            |+++++++++++++++++++++++++++++++++++++++++++++++   | 94% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++++  | 95% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++++  | 96% ~00s            |+++++++++++++++++++++++++++++++++++++++++++++++++ | 97% ~00s            |+++++++++++++++++++++++++++++++++++++++++++++++++ | 98% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++++++| 99% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++++++| 100% elapsed=02s  
#> Test heterogeneity change at protein
#>   |                                                  | 0 % ~calculating    |+                                                 | 1 % ~01s            |++                                                | 2 % ~01s            |++                                                | 3 % ~01s            |+++                                               | 4 % ~01s            |+++                                               | 5 % ~01s            |++++                                              | 6 % ~01s            |++++                                              | 7 % ~01s            |+++++                                             | 8 % ~01s            |+++++                                             | 9 % ~01s            |++++++                                            | 10% ~01s            |++++++                                            | 11% ~01s            |+++++++                                           | 12% ~01s            |+++++++                                           | 13% ~00s            |++++++++                                          | 14% ~00s            |++++++++                                          | 15% ~00s            |+++++++++                                         | 16% ~00s            |+++++++++                                         | 17% ~00s            |++++++++++                                        | 18% ~00s            |++++++++++                                        | 19% ~00s            |+++++++++++                                       | 20% ~00s            |+++++++++++                                       | 21% ~00s            |++++++++++++                                      | 22% ~00s            |++++++++++++                                      | 23% ~00s            |+++++++++++++                                     | 24% ~00s            |+++++++++++++                                     | 25% ~00s            |++++++++++++++                                    | 26% ~00s            |++++++++++++++                                    | 27% ~00s            |+++++++++++++++                                   | 28% ~00s            |+++++++++++++++                                   | 29% ~00s            |++++++++++++++++                                  | 30% ~00s            |++++++++++++++++                                  | 31% ~00s            |+++++++++++++++++                                 | 32% ~00s            |+++++++++++++++++                                 | 33% ~00s            |++++++++++++++++++                                | 34% ~00s            |++++++++++++++++++                                | 35% ~00s            |+++++++++++++++++++                               | 36% ~00s            |+++++++++++++++++++                               | 37% ~00s            |++++++++++++++++++++                              | 38% ~00s            |++++++++++++++++++++                              | 39% ~00s            |+++++++++++++++++++++                             | 40% ~00s            |+++++++++++++++++++++                             | 41% ~00s            |++++++++++++++++++++++                            | 42% ~00s            |++++++++++++++++++++++                            | 43% ~00s            |+++++++++++++++++++++++                           | 44% ~00s            |+++++++++++++++++++++++                           | 45% ~00s            |++++++++++++++++++++++++                          | 46% ~00s            |++++++++++++++++++++++++                          | 47% ~00s            |+++++++++++++++++++++++++                         | 48% ~00s            |+++++++++++++++++++++++++                         | 49% ~00s            |++++++++++++++++++++++++++                        | 51% ~00s            |++++++++++++++++++++++++++                        | 52% ~00s            |+++++++++++++++++++++++++++                       | 53% ~00s            |+++++++++++++++++++++++++++                       | 54% ~00s            |++++++++++++++++++++++++++++                      | 55% ~00s            |++++++++++++++++++++++++++++                      | 56% ~00s            |+++++++++++++++++++++++++++++                     | 57% ~00s            |+++++++++++++++++++++++++++++                     | 58% ~00s            |++++++++++++++++++++++++++++++                    | 59% ~00s            |++++++++++++++++++++++++++++++                    | 60% ~00s            |+++++++++++++++++++++++++++++++                   | 61% ~00s            |+++++++++++++++++++++++++++++++                   | 62% ~00s            |++++++++++++++++++++++++++++++++                  | 63% ~00s            |++++++++++++++++++++++++++++++++                  | 64% ~00s            |+++++++++++++++++++++++++++++++++                 | 65% ~00s            |+++++++++++++++++++++++++++++++++                 | 66% ~00s            |++++++++++++++++++++++++++++++++++                | 67% ~00s            |++++++++++++++++++++++++++++++++++                | 68% ~00s            |+++++++++++++++++++++++++++++++++++               | 69% ~00s            |+++++++++++++++++++++++++++++++++++               | 70% ~00s            |++++++++++++++++++++++++++++++++++++              | 71% ~00s            |++++++++++++++++++++++++++++++++++++              | 72% ~00s            |+++++++++++++++++++++++++++++++++++++             | 73% ~00s            |+++++++++++++++++++++++++++++++++++++             | 74% ~00s            |++++++++++++++++++++++++++++++++++++++            | 75% ~00s            |++++++++++++++++++++++++++++++++++++++            | 76% ~00s            |+++++++++++++++++++++++++++++++++++++++           | 77% ~00s            |+++++++++++++++++++++++++++++++++++++++           | 78% ~00s            |++++++++++++++++++++++++++++++++++++++++          | 79% ~00s            |++++++++++++++++++++++++++++++++++++++++          | 80% ~00s            |+++++++++++++++++++++++++++++++++++++++++         | 81% ~00s            |+++++++++++++++++++++++++++++++++++++++++         | 82% ~00s            |++++++++++++++++++++++++++++++++++++++++++        | 83% ~00s            |++++++++++++++++++++++++++++++++++++++++++        | 84% ~00s            |+++++++++++++++++++++++++++++++++++++++++++       | 85% ~00s            |+++++++++++++++++++++++++++++++++++++++++++       | 86% ~00s            |++++++++++++++++++++++++++++++++++++++++++++      | 87% ~00s            |++++++++++++++++++++++++++++++++++++++++++++      | 88% ~00s            |+++++++++++++++++++++++++++++++++++++++++++++     | 89% ~00s            |+++++++++++++++++++++++++++++++++++++++++++++     | 90% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++    | 91% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++    | 92% ~00s            |+++++++++++++++++++++++++++++++++++++++++++++++   | 93% ~00s            |+++++++++++++++++++++++++++++++++++++++++++++++   | 94% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++++  | 95% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++++  | 96% ~00s            |+++++++++++++++++++++++++++++++++++++++++++++++++ | 97% ~00s            |+++++++++++++++++++++++++++++++++++++++++++++++++ | 98% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++++++| 99% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++++++| 100% elapsed=01s  

head(res)
#>        trait score_type                                    feature group1
#> 1 GlycanSize         mu                        VPGJQTSTTLK (L1CAM) Normal
#> 2 GlycanSize         mu                        JSTKEEILAALEK (SAP) Normal
#> 3 GlycanSize         mu                         FJHTQTIQQK (NRCAM) Normal
#> 4 GlycanSize         mu                    GELJTSIFSSRPIDK (PCYOX) Normal
#> 5 GlycanSize         mu                       JHSIFLADINQER (PPT1) Normal
#> 6 GlycanSize         mu VAWLJR (CEPU1|IGLO5|LSAMP|NTRI|OBCAM|OPCM) Normal
#>        group2 mean_group1 mean_group2        diff      pval n_group1 n_group2
#> 1 Symptomatic    7.794117    7.895833 -0.10171633 0.5000000        8        4
#> 2 Symptomatic    7.014646    7.075000 -0.06035354 0.1666667        9        7
#> 3 Symptomatic    7.310174    7.388095 -0.07792158 0.3333333        8        7
#> 4 Symptomatic    8.662879    8.750000 -0.08712121 0.8333333        4        3
#> 5 Symptomatic    7.447484    7.527690 -0.08020571 0.6666667        8        7
#> 6 Symptomatic    8.104233    8.190861 -0.08662772 0.1666667        9        8
#>        method level
#> 1 permutation  site
#> 2 permutation  site
#> 3 permutation  site
#> 4 permutation  site
#> 5 permutation  site
#> 6 permutation  site